requirements.txt
polygon
pandas
matplotlib
seaborn
vaderSentiment
scipy
python-dotenv